--- src/PSSVoidFunctionCall.java
@@ -1,17 +1,0 @@
-import java.util.*;
-
-public class PSSVoidFunctionCall extends PSSProcStmt {
-
-	PSSFunctionCall func = null;
-
-	PSSVoidFunctionCall(PSSFunctionCall func) {
-		this.func = func;
-	}
-
-	public void eval(PSSInst var) {
-		if (func != null)
-			func.eval(var);
-	}
-
-}
-
